requirements.txt
fastapi[all]
asyncio
asyncpg
python-dotenv
jinja2
